refactor: sepaate router logic
